+++ shibuya-core-bench/bench/Bench/Handler.hs
@@ -131,6 +131,7 @@ createIngestedMessages n = do
                 partition = Nothing,
                 enqueuedAt = Just now,
                 traceContext = Nothing,
+                headers = Nothing,
                 attempt = Nothing,
                 attributes = HashMap.empty,
                 payload = BenchMessage i (Text.pack $ "payload-" <> show i)